Fixed bug when quality for a type was set to zero
If the quality was set to zero (0 or 0.0), then the logic of the
original code was defaulting it to a value of one (1). This commit
ensures that, if the Accept header sets the quality value to zero, the
wishes of the client are kept intact.

In addition, if the Accept header set the quality value on a media
type to zero, the fitness value was still being calculating fairly high
-- even higher than other supported media types that matched the */*
type. As a result, the media type set to zero was appearing at the
bottom of the list after sorting, and it was being selected as the best
match. This commit unsets values that have a quality of zero so that
they will not be selected for the best match.